refactor: add the box subtitle component to which you apply the pattern composition
