+++ src/components/boxSubtitle/components/containerSubtitle/index.tsx
@@ -1,0 +1,9 @@
+import * as S from './style';
+
+export default function ContainerSubtitle({
+    children,
+}: {
+    children: React.ReactNode;
+}) {
+    return <S.BoxSubtitle>{children}</S.BoxSubtitle>;
+}
