+++ src/components/boxSubtitle/components/containerSubtitle/style.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const BoxSubtitle = styled.div`
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+`;
