+++ src/components/boxSubtitle/index.tsx
@@ -1,0 +1,11 @@
+import Category from './components/categoryName';
+import Container from './components/containerSubtitle';
+import Subtitle from './components/subtitle';
+
+const BoxSubtitle = {
+    Container,
+    Category,
+    Subtitle,
+};
+
+export default BoxSubtitle;
